pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gai-mcp"
version = "0.1.0"
description = "AI 自动游玩游戏的 MCP 工具"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0",
    "pywin32>=306",
    "pyvda>=0.4.0",
    "mss>=9.0",
    "Pillow>=10.0",
    "pyautogui>=0.9",
    "anthropic>=0.40",
    "openai>=1.50",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "pyperclip>=1.8",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[project.scripts]
gai-mcp = "gai_mcp.server:main"
gai-mcp-web = "gai_mcp.web.run:main"

[tool.hatch.build.targets.wheel]
packages = ["src/gai_mcp"]
